Make the code much simpler
Much of the code can be removed and still function the same. The main
thrust of this change is that there doesn't need to be an "each()" loop
for iterating over the jQuery collection, because setting all their text
with one call to "text()" works just the same. This removed the
necessity of having specific "linked" and "text" vars, since we can just
call this.val() directly.